refactor(sharing-dialog): flatten tab hierarchy
