+++ components/sharing-dialog/src/tabs/index.js
@@ -1,0 +1,1 @@
+export { TabbedContent } from './tabbed-content.js'
